message-database/messenger/requirements.txt
requests==2.27.1
confluent-kafka==1.8.2
pymongo==4.1.1
mysql-connector-python==8.0.29